Calculate the Area Under the Disease Progress Curve (AUDPC)

This is a modified version of agricolae::audpc(), which checks for NA values and will return NA values if specified in the argument. See help(agricolae::audpc()) for more details. Note that this function starts with a capital “A” to avoid namespace clashes with agricolae.

Usage

Audpc(evaluation, dates, type = "absolute", na.rm = FALSE)

Arguments

evaluation

Table of data of the evaluations: Data frame

dates

Vector of dates corresponding to each evaluation

type

relative, absolute

na.rm

weather to remove NA values or return NA when there are only NA values

Value

Vector with relative or absolute AUDPC.

Examples

# see examples in help(agricolae::audpc)
dates <- c(14, 21, 28) # days
# example 1: evaluation - vector
evaluation <- c(40, 80, 90)
Audpc(evaluation, dates)
#> evaluation 
#>       1015 
dates <- c(14, 21, NA) # days
Audpc(evaluation, dates)
#> [1] NA
# agricolae::audpc(evaluation,dates)
dates <- c(14, 21, 28) # days
evaluation <- c(40, 80, NA)
Audpc(evaluation, dates)
#> [1] NA
# agricolae::audpc(evaluation,dates)