Calculate estimated weather from coefficients

Calculate estimated weather from coefficients

Usage

calc_estimated_weather(
  w,
  start_date = "2023-04-01",
  end_date = "2023-11-30",
  lat,
  lon,
  n_stations = 1:4,
  na.rm = FALSE
)

Arguments

w

A data.table of weather coefficients, output from epiphytoolR::get_weather_coefs

start_date

a character string of a date value indicating the first date for in the returning data.table Must be in ISO8601 format (YYYY-MM-DD), e.g. “2020-04-26”

end_date

a character string of a date value indicating the last date for in the returning data.table Must be in ISO8601 format (YYYY-MM-DD), e.g. “2020-04-26”

lat

numeric, latitude of the query coordinates where weather should be estimated. If missing all stations will be returned

lon

numeric, longitude of the query coordinates where weather should be estimated. If missing all stations will be returned

n_stations

integer or vector of integers indicating the number of station/s to return from the closest (1), or 3rd closest (3) or closest five stations (1:5). ect

na.rm

logical, remove all weather data from stations with NA rain_fall frequency Not Recommended. We advise the best way is to manually remove weather stations with NAs or correct the weather data. This argument is available if these two options are not available to the user. Default is FALSE

Value

A data.table output of calculated on get_weather_coefs.R with the following columns: station - Weather station name; lat - latitude; lon - longitude; rh - NA currently not supported see epiphytoolR github issue #14; yearday - integer, day of the year, see data.table::yday(); wd_rd - numeric, mean wind direction from raw data; wd_sd_rd - numeric, standard deviation of wind direction from raw data; ws_rd - numeric, mean wind speed from raw data; ws_sd_rd - numeric, standard deviation of wind speed from raw data; rain_freq - numeric, proportional chance of rainfall on this dat 0 - 1

can be formated with format_weather()

Examples

dat <- data.frame(
  station_name = "w_STATION",
  lat = -runif(1, 15.5, 28),
  lon = runif(1, 115, 150),
  state = "SA",
  yearday = 1:365,
  wd_rw = abs(rnorm(365, 180, 90)),
  wd_sd_rw = rnorm(365, 80, 20),
  ws_rw = runif(365, 1, 60),
  ws_sd_rw = abs(rnorm(365, 10, sd = 5)),
  rain_freq = runif(365, 0.05, 0.45)
  )

  calc_estimated_weather(w = dat,
    lat = -25,
    lon = 130,
    n_stations = 1)
#> Warning: 'max_temp' and 'min_temp' not detected, returning NAs for mean daily 'temp'
#>                     times   station yearday       lat      lon  state    wd_rw
#>                    <POSc>    <char>   <int>     <num>    <num> <char>    <num>
#>    1: 2023-04-01 00:00:00 w_STATION      91 -16.50938 144.2017     SA 275.6792
#>    2: 2023-04-01 01:00:00 w_STATION      91 -16.50938 144.2017     SA 275.6792
#>    3: 2023-04-01 02:00:00 w_STATION      91 -16.50938 144.2017     SA 275.6792
#>    4: 2023-04-01 03:00:00 w_STATION      91 -16.50938 144.2017     SA 275.6792
#>    5: 2023-04-01 04:00:00 w_STATION      91 -16.50938 144.2017     SA 275.6792
#>   ---                                                                         
#> 5829: 2023-11-29 20:00:00 w_STATION     333 -16.50938 144.2017     SA 164.7648
#> 5830: 2023-11-29 21:00:00 w_STATION     333 -16.50938 144.2017     SA 164.7648
#> 5831: 2023-11-29 22:00:00 w_STATION     333 -16.50938 144.2017     SA 164.7648
#> 5832: 2023-11-29 23:00:00 w_STATION     333 -16.50938 144.2017     SA 164.7648
#> 5833: 2023-11-30 00:00:00 w_STATION     334 -16.50938 144.2017     SA 206.5527
#>       wd_sd_rw     ws_rw  ws_sd_rw rain_freq distance    rh  rain  temp
#>          <num>     <num>     <num>     <num>    <num> <num> <int> <num>
#>    1: 80.07136 10.924185 13.696844 0.4303483 1752.619    NA     1    NA
#>    2: 80.07136 10.924185 13.696844 0.4303483 1752.619    NA     1    NA
#>    3: 80.07136 10.924185 13.696844 0.4303483 1752.619    NA     1    NA
#>    4: 80.07136 10.924185 13.696844 0.4303483 1752.619    NA     1    NA
#>    5: 80.07136 10.924185 13.696844 0.4303483 1752.619    NA     1    NA
#>   ---                                                                  
#> 5829: 91.00031  6.263042  3.451388 0.4428285 1752.619    NA     0    NA
#> 5830: 91.00031  6.263042  3.451388 0.4428285 1752.619    NA     0    NA
#> 5831: 91.00031  6.263042  3.451388 0.4428285 1752.619    NA     0    NA
#> 5832: 91.00031  6.263042  3.451388 0.4428285 1752.619    NA     0    NA
#> 5833: 80.56714 54.921931  6.464564 0.2794979 1752.619    NA     0    NA
#>              ws       wd    wd_sd  YYYY    MM    DD    hh    mm
#>           <num>    <num>    <num> <int> <int> <int> <int> <int>
#>    1: 27.922806 275.6792 80.07136  2023     4     1     0     0
#>    2: 27.922806 275.6792 80.07136  2023     4     1     1     0
#>    3: 27.922806 275.6792 80.07136  2023     4     1     2     0
#>    4: 27.922806 275.6792 80.07136  2023     4     1     3     0
#>    5: 27.922806 275.6792 80.07136  2023     4     1     4     0
#>   ---                                                          
#> 5829:  7.706564 164.7648 91.00031  2023    11    29    20     0
#> 5830:  7.706564 164.7648 91.00031  2023    11    29    21     0
#> 5831:  7.706564 164.7648 91.00031  2023    11    29    22     0
#> 5832:  7.706564 164.7648 91.00031  2023    11    29    23     0
#> 5833: 52.073548 206.5527 80.56714  2023    11    30     0     0