Merge BOM axf weather data

@details This function takes new BOM axf files which hold 72 hours of weather observations

Usage

merge_axf_weather(
  File_compressed,
  File_axf = "IDQ60910.99123.axf",
  File_formatted = "NTamborine.csv",
  base_dir = getwd(),
  verbose = FALSE
)

Arguments

File_compressed

character, file path of compressed weather file "tgz"

File_axf

character, filename of axf weather data observation file from bom

File_formatted

character, filename and path to the formated file which has previously merged data

base_dir

character file path giving the base directory for file_formatted

verbose

logical print extra messages to asssist debugging

Value

data.table, of merged dataset